requirements.txt
# Django & REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# API Documentation
drf-yasg>=1.21,<2.0

# Database
psycopg2-binary>=2.9,<3.0

# Filtering & Pagination
django-filter>=23.5,<24.0

# Environment variables
python-decouple>=3.8,<4.0

# Production server & static files
gunicorn>=21.0,<22.0
whitenoise>=6.6,<7.0
